requirements.txt
Flask==0.10.1
Flask-RESTful==0.2.12
Flask-Sockets==0.1
Jinja2==2.7.3
MarkupSafe==0.23
Werkzeug==0.9.6
aniso8601==0.83
gevent==1.0.1
gevent-websocket==0.9.3
greenlet==0.4.4
gunicorn==19.1.1
itsdangerous==0.24
pytz==2014.7
redis==2.10.3
six==1.8.0
git+git://github.com/livni/redis_wrap.git@1fbdc2f779785a88ba43fd209fea6573305544ed#egg=redis_wrap
#ws4py==0.3.4
websocket-client>=0.14.1
